+++ gms-server/src/main/java/org/gms/dao/entity/TrocklocationsDO.java
@@ -28,7 +28,7 @@ public class TrocklocationsDO implements Serializable {
     private static final long serialVersionUID = 1L;
 
     @Id(keyType = KeyType.Auto)
-    private Integer trockid;
+    private Long trockid;
 
     private Integer characterid;
